Fixed issue with nil, and removed more tests of the old version
