+++ common/authorization/authorizer_test.go
@@ -26,6 +26,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/uber/cadence/common/types"
 
 	"github.com/uber/cadence/common"
 )
@@ -94,4 +95,90 @@ func Test_validatePermission(t *testing.T) {
 			tt.wantErr(t, validatePermission(tt.claims, tt.attributes, tt.data))
 		})
 	}
+}
+
+func TestSignalWithStartWorkflowExecutionRequestSerializeForLogging(t *testing.T) {
+	tests := map[string]struct {
+		input               *types.SignalWithStartWorkflowExecutionRequest
+		expectedOutput      string
+		expectedErrorOutput error
+	}{
+		"complete request without error": {
+			input:               createNewSignalWithStartWorkflowExecutionRequest(),
+			expectedOutput:      "{\"domain\":\"testDomain\",\"workflowId\":\"testWorkflowID\",\"workflowType\":{\"name\":\"testWorkflowType\"},\"taskList\":{\"name\":\"testTaskList\",\"kind\":\"STICKY\"},\"executionStartToCloseTimeoutSeconds\":1,\"taskStartToCloseTimeoutSeconds\":1,\"identity\":\"testIdentity\",\"requestId\":\"DF66E35D-A5B0-425D-8731-6AAC4A4B6368\",\"workflowIdReusePolicy\":\"AllowDuplicate\",\"signalName\":\"testRequest\",\"control\":\"dGVzdENvbnRyb2w=\",\"retryPolicy\":{\"initialIntervalInSeconds\":1,\"backoffCoefficient\":1,\"maximumIntervalInSeconds\":1,\"maximumAttempts\":1,\"nonRetriableErrorReasons\":[\"testArray\"],\"expirationIntervalInSeconds\":1},\"cronSchedule\":\"testSchedule\",\"header\":{},\"delayStartSeconds\":1,\"jitterStartSeconds\":1,\"firstRunAtTimestamp\":1}",
+			expectedErrorOutput: nil,
+		},
+
+		"empty request without error": {
+			input:               &types.SignalWithStartWorkflowExecutionRequest{},
+			expectedOutput:      "{}",
+			expectedErrorOutput: nil,
+		},
+
+		"nil request without error": {
+			input:               nil,
+			expectedOutput:      "",
+			expectedErrorOutput: nil,
+		},
+	}
+
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			assert.NotPanics(t, func() {
+				wrappedInput := NewFilteredRequestBody(test.input)
+				output, err := wrappedInput.SerializeForLogging()
+				assert.Equal(t, test.expectedOutput, output)
+				assert.Equal(t, test.expectedErrorOutput, err)
+				assert.NotContains(t, output, "PII")
+			})
+		})
+	}
+}
+
+func createNewSignalWithStartWorkflowExecutionRequest() *types.SignalWithStartWorkflowExecutionRequest {
+	testTasklistKind := types.TaskListKind(1)
+	testExecutionStartToCloseTimeoutSeconds := int32(1)
+	testTaskStartToCloseTimeoutSeconds := int32(1)
+	testWorkflowIDReusePolicy := types.WorkflowIDReusePolicy(1)
+	testDelayStartSeconds := int32(1)
+	testJitterStartSeconds := int32(1)
+	testFirstRunAtTimestamp := int64(1)
+	piiTestArray := []byte("testInputPII")
+	piiTestMap := make(map[string][]byte)
+	piiTestMap["PII"] = piiTestArray
+
+	testReq := &types.SignalWithStartWorkflowExecutionRequest{
+		Domain:       "testDomain",
+		WorkflowID:   "testWorkflowID",
+		WorkflowType: &types.WorkflowType{Name: "testWorkflowType"},
+		TaskList: &types.TaskList{
+			Name: "testTaskList",
+			Kind: &testTasklistKind,
+		},
+		Input:                               piiTestArray,
+		ExecutionStartToCloseTimeoutSeconds: &testExecutionStartToCloseTimeoutSeconds,
+		TaskStartToCloseTimeoutSeconds:      &testTaskStartToCloseTimeoutSeconds,
+		Identity:                            "testIdentity",
+		RequestID:                           "DF66E35D-A5B0-425D-8731-6AAC4A4B6368",
+		WorkflowIDReusePolicy:               &testWorkflowIDReusePolicy,
+		SignalName:                          "testRequest",
+		SignalInput:                         piiTestArray,
+		Control:                             []byte("testControl"),
+		RetryPolicy: &types.RetryPolicy{
+			InitialIntervalInSeconds:    1,
+			BackoffCoefficient:          1,
+			MaximumIntervalInSeconds:    1,
+			MaximumAttempts:             1,
+			NonRetriableErrorReasons:    []string{"testArray"},
+			ExpirationIntervalInSeconds: 1,
+		},
+		CronSchedule:        "testSchedule",
+		Memo:                &types.Memo{Fields: piiTestMap},
+		SearchAttributes:    &types.SearchAttributes{IndexedFields: piiTestMap},
+		Header:              &types.Header{Fields: map[string][]byte{}},
+		DelayStartSeconds:   &testDelayStartSeconds,
+		JitterStartSeconds:  &testJitterStartSeconds,
+		FirstRunAtTimestamp: &testFirstRunAtTimestamp,
+	}
+	return testReq
 }
