--- common/types_test/shared_test.go
@@ -1,188 +1,0 @@
-// The MIT License (MIT)
-
-// Copyright (c) 2017-2020 Uber Technologies Inc.
-
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-package types_test
-
-import (
-	"testing"
-	"unsafe"
-
-	"github.com/stretchr/testify/assert"
-
-	"github.com/uber/cadence/common/authorization"
-	"github.com/uber/cadence/common/types"
-)
-
-func TestSignalWithStartWorkflowExecutionRequestSerializeForLogging(t *testing.T) {
-	tests := map[string]struct {
-		input               *types.SignalWithStartWorkflowExecutionRequest
-		expectedOutput      string
-		expectedErrorOutput error
-	}{
-		"complete request without error": {
-			input:               createNewSignalWithStartWorkflowExecutionRequest(),
-			expectedOutput:      "{\"domain\":\"testDomain\",\"workflowId\":\"testWorkflowID\",\"workflowType\":{\"name\":\"testWorkflowType\"},\"taskList\":{\"name\":\"testTaskList\",\"kind\":\"STICKY\"},\"executionStartToCloseTimeoutSeconds\":1,\"taskStartToCloseTimeoutSeconds\":1,\"identity\":\"testIdentity\",\"requestId\":\"DF66E35D-A5B0-425D-8731-6AAC4A4B6368\",\"workflowIdReusePolicy\":\"AllowDuplicate\",\"signalName\":\"testRequest\",\"control\":\"dGVzdENvbnRyb2w=\",\"retryPolicy\":{\"initialIntervalInSeconds\":1,\"backoffCoefficient\":1,\"maximumIntervalInSeconds\":1,\"maximumAttempts\":1,\"nonRetriableErrorReasons\":[\"testArray\"],\"expirationIntervalInSeconds\":1},\"cronSchedule\":\"testSchedule\",\"header\":{},\"delayStartSeconds\":1,\"jitterStartSeconds\":1,\"firstRunAtTimestamp\":1}",
-			expectedErrorOutput: nil,
-		},
-
-		"empty request without error": {
-			input:               &types.SignalWithStartWorkflowExecutionRequest{},
-			expectedOutput:      "{}",
-			expectedErrorOutput: nil,
-		},
-
-		"nil request without error": {
-			input:               nil,
-			expectedOutput:      "",
-			expectedErrorOutput: nil,
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			assert.NotPanics(t, func() {
-				wrappedInput := authorization.NewFilteredRequestBody(test.input)
-				output, err := wrappedInput.SerializeForLogging()
-				assert.Equal(t, test.expectedOutput, output)
-				assert.Equal(t, test.expectedErrorOutput, err)
-				assert.NotContains(t, output, "PII")
-			})
-		})
-	}
-}
-
-func TestSerializeRequest(t *testing.T) {
-	// test serializing a normal request
-	testReq := createNewSignalWithStartWorkflowExecutionRequest()
-	serializeRes, err := types.SerializeRequest(testReq)
-
-	expectRes := "{\"domain\":\"testDomain\",\"workflowId\":\"testWorkflowID\",\"workflowType\":{\"name\":\"testWorkflowType\"}," +
-		"\"taskList\":{\"name\":\"testTaskList\",\"kind\":\"STICKY\"},\"executionStartToCloseTimeoutSeconds\":1," +
-		"\"taskStartToCloseTimeoutSeconds\":1,\"identity\":\"testIdentity\",\"requestId\":\"DF66E35D-A5B0-425D-8731-6AAC4A4B6368\"," +
-		"\"workflowIdReusePolicy\":\"AllowDuplicate\",\"signalName\":\"testRequest\",\"control\":\"dGVzdENvbnRyb2w=\"," +
-		"\"retryPolicy\":{\"initialIntervalInSeconds\":1,\"backoffCoefficient\":1,\"maximumIntervalInSeconds\":1," +
-		"\"maximumAttempts\":1,\"nonRetriableErrorReasons\":[\"testArray\"],\"expirationIntervalInSeconds\":1}," +
-		"\"cronSchedule\":\"testSchedule\",\"header\":{},\"delayStartSeconds\":1,\"jitterStartSeconds\":1,\"firstRunAtTimestamp\":1}"
-	expectErr := error(nil)
-
-	assert.Equal(t, expectRes, serializeRes)
-	assert.Equal(t, expectErr, err)
-
-	assert.NotPanics(t, func() {
-		types.SerializeRequest(nil)
-	})
-
-}
-
-func createNewSignalWithStartWorkflowExecutionRequest() *types.SignalWithStartWorkflowExecutionRequest {
-	testTasklistKind := types.TaskListKind(1)
-	testExecutionStartToCloseTimeoutSeconds := int32(1)
-	testTaskStartToCloseTimeoutSeconds := int32(1)
-	testWorkflowIDReusePolicy := types.WorkflowIDReusePolicy(1)
-	testDelayStartSeconds := int32(1)
-	testJitterStartSeconds := int32(1)
-	testFirstRunAtTimestamp := int64(1)
-	piiTestArray := []byte("testInputPII")
-	piiTestMap := make(map[string][]byte)
-	piiTestMap["PII"] = piiTestArray
-
-	testReq := &types.SignalWithStartWorkflowExecutionRequest{
-		Domain:       "testDomain",
-		WorkflowID:   "testWorkflowID",
-		WorkflowType: &types.WorkflowType{Name: "testWorkflowType"},
-		TaskList: &types.TaskList{
-			Name: "testTaskList",
-			Kind: &testTasklistKind,
-		},
-		Input:                               piiTestArray,
-		ExecutionStartToCloseTimeoutSeconds: &testExecutionStartToCloseTimeoutSeconds,
-		TaskStartToCloseTimeoutSeconds:      &testTaskStartToCloseTimeoutSeconds,
-		Identity:                            "testIdentity",
-		RequestID:                           "DF66E35D-A5B0-425D-8731-6AAC4A4B6368",
-		WorkflowIDReusePolicy:               &testWorkflowIDReusePolicy,
-		SignalName:                          "testRequest",
-		SignalInput:                         piiTestArray,
-		Control:                             []byte("testControl"),
-		RetryPolicy: &types.RetryPolicy{
-			InitialIntervalInSeconds:    1,
-			BackoffCoefficient:          1,
-			MaximumIntervalInSeconds:    1,
-			MaximumAttempts:             1,
-			NonRetriableErrorReasons:    []string{"testArray"},
-			ExpirationIntervalInSeconds: 1,
-		},
-		CronSchedule:        "testSchedule",
-		Memo:                &types.Memo{Fields: piiTestMap},
-		SearchAttributes:    &types.SearchAttributes{IndexedFields: piiTestMap},
-		Header:              &types.Header{Fields: map[string][]byte{}},
-		DelayStartSeconds:   &testDelayStartSeconds,
-		JitterStartSeconds:  &testJitterStartSeconds,
-		FirstRunAtTimestamp: &testFirstRunAtTimestamp,
-	}
-	return testReq
-}
-
-func TestDataBlobDeepCopy(t *testing.T) {
-	tests := []struct {
-		name  string
-		input *types.DataBlob
-	}{
-		{
-			name:  "nil",
-			input: nil,
-		},
-		{
-			name:  "empty",
-			input: &types.DataBlob{},
-		},
-		{
-			name: "thrift ok",
-			input: &types.DataBlob{
-				EncodingType: types.EncodingTypeThriftRW.Ptr(),
-				Data:         []byte("some thrift data"),
-			},
-		},
-		{
-			name: "json ok",
-			input: &types.DataBlob{
-				EncodingType: types.EncodingTypeJSON.Ptr(),
-				Data:         []byte("some json data"),
-			},
-		},
-	}
-
-	for _, tc := range tests {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			got := tc.input.DeepCopy()
-			assert.Equal(t, tc.input, got)
-			if tc.input != nil && tc.input.Data != nil && identicalByteArray(tc.input.Data, got.Data) {
-				t.Error("expected DeepCopy to return a new data slice")
-			}
-		})
-	}
-}
-
-// identicalByteArray returns true if a and b are the same slice, false otherwise.
-func identicalByteArray(a, b []byte) bool {
-	return len(a) == len(b) && unsafe.SliceData(a) == unsafe.SliceData(b)
-}
